Do not fail, if the caller didn't pick up response from previous call
During our testing of a new CAC driver in OpenSC, with this library, we
encountered a problem with |libcacard| failing and the driver returning
only a fraction of the requested objects.

The problem is that the Emulator wants to return the data (properly
signalized by 61 (RESPONSE BYTES) in SW1), but this is ignored for some
reason in some of our calls from OpenSC. The Emulator should not fail
hard for the next independent request, rather silently drop the buffer
and serve the ongoing APDU request (I would left for consideration to
somehow log such problem).

Patch was successfully tested on Fedora 24 host and solves our problem
(though we worked around the problem already in the driver too).

Reviewed-by: Marc-André Lureau <marcandre.lureau@redhat.com>