tests: Fix 'simpletlv' on i686 arches
The simpletlv_* functions return lengths as an int, but the test is
casting it to a size_t before comparing it against -1. This is not
giving the expected result on i686, causing the test to fail. The
variables used in the test do not need to be size_t, and can be changed
to int to match the return value and avoid the failure.

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>
Acked-by: Marc-André Lureau <marcandre.lureau@redhat.com>