vcard_emul_nss: Drop the key caching to simplify error handling
It could happen with PKCS#11 modules that (correctly) invalidate object
handles after logout (which was introduced in 0d3a683a), that the handles
are not valid when we try to use the objects again.

This is trying to address this use case, which I noticed was breaking
CI with SoftHSM PKCS#11 modules.

Signed-off-by: Jakub Jelen <jjelen@redhat.com>