Add first simple fuzzer
Per [1], we are good to go to integrate in oss-fuzz. This adds
aa simple fuzzer and some housekeeping/preparatory work around
build system to be able to build fuzzers for oss-fuzz.

The fuzzers can be built only with new meson build system as
trying to make it working with autotools was quite a pain.

The oss-fuzz is already running on my fork temporarily, before
this will be merged as in [2].

[1] https://github.com/google/oss-fuzz/pull/3019
[2] https://github.com/google/oss-fuzz/pull/3466

Signed-off-by: Jakub Jelen <jjelen@redhat.com>
Acked-by: Frediano Ziglio <fziglio@redhat.com>